Seperated Reducer logic to reducer.tsx
