+++ src/components/Form.tsx
@@ -7,14 +7,14 @@ import {
   formField,
   formData,
   textFieldTypes,
-  radioField,
   formKinds,
 } from "../types/formTypes";
 import LabeledDropdown from "./LabeledDropdown";
 import LabeledRadio from "./LabeledRadio";
 import LabeledMultiSelect from "./LabeledMultiSelect";
+import { reducerForm } from "../reducer/reducers";
 
-const formFields: formField[] = [
+export const formFields: formField[] = [
   {
     kind: "text",
     id: 1,
@@ -52,284 +52,14 @@ const formFields: formField[] = [
   },
 ];
 
-type RemoveAction = {
-  type: "remove_field";
-  id: number;
-};
-
-type AddAction = {
-  type: "add_field";
-  kind: formKinds;
-  label: string;
-};
-
-type AddOption = {
-  type: "add_option";
-  id: number;
-  str: string;
-};
-
-type RemoveOption = {
-  type: "remove_option";
-  id: number;
-  formID: number;
-};
-
-type UpdateTitleAction = {
-  type: "update_title";
-  title: string;
-};
-
-type UpdateLabel = {
-  type: "update_label";
-  id: number;
-  str: string;
-};
-
-type UpdateRadio = {
-  type: "update_radio";
-  id: number;
-  stOption: string;
-  formID: number;
-};
-
-type SaveForm = {
-  type: "save_form";
-};
-
-type FormAction =
-  | AddAction
-  | RemoveAction
-  | UpdateTitleAction
-  | AddOption
-  | RemoveOption
-  | UpdateLabel
-  | UpdateRadio
-  | SaveForm;
-
-const getAllOptions = (
-  oldOptions: { id: number; option: string }[],
-  option: string
-) => {
-  return [...oldOptions, { id: Number(new Date()), option: option }];
-};
-
 export default function Form(props: { formId: number }) {
   //Reducer.
-  const reducer: (state: formData, action: FormAction) => formData = (
-    state: formData,
-    action: FormAction
-  ) => {
-    switch (action.type) {
-      case "add_field": {
-        console.log(action.kind + "kind");
-        let newField;
-        if (action.kind === "multi-select") {
-          newField = getNewField("multi-select", action.label);
-        } else {
-          newField = getNewField(action.kind, action.label);
-        }
-        console.log(newField);
-        if (newField.label.length > 0)
-          return {
-            ...state,
-            formFields: [...state.formFields, newField],
-          };
-        return state;
-      }
-      case "remove_field": {
-        return {
-          ...state,
-          formFields: state.formFields.filter(
-            (field) => field.id !== action.id
-          ),
-        };
-      }
-      case "update_title": {
-        return {
-          ...state,
-          title: action.title,
-        };
-      }
-      case "add_option": {
-        let prevState = state.formFields.find((ele) => ele.id === action.id);
-        let oldOption: { id: number; option: string }[] = [];
-        if (prevState && "options" in prevState) {
-          oldOption = [...prevState.options];
-        }
-        const allOptions = getAllOptions(oldOption, action.str);
-        return {
-          ...state,
-          formFields: state.formFields.map((ele) => {
-            if (
-              ele.id === action.id &&
-              (ele.kind === "dropdown" ||
-                ele.kind === "radio" ||
-                ele.kind === "multiple")
-            ) {
-              return {
-                ...ele,
-                options: allOptions,
-              };
-            } else {
-              return ele;
-            }
-          }),
-        };
-      }
-      case "remove_option": {
-        const formfields = state.formFields.filter(
-          (field) => field.id === action.formID
-        )[0] as radioField;
-
-        return {
-          ...state,
-          formFields: state.formFields.map((ele) => {
-            if (
-              ele.id === action.formID &&
-              (ele.kind === "dropdown" ||
-                ele.kind === "radio" ||
-                ele.kind === "multiple")
-            ) {
-              return {
-                ...ele,
-                options: formfields.options.filter(
-                  (ele) => ele.id !== action.id
-                ),
-              };
-            } else {
-              return ele;
-            }
-          }),
-        };
-      }
-      case "update_label": {
-        return {
-          ...state,
-          formFields: state.formFields.map((ele) => {
-            if (ele.id === action.id) {
-              return {
-                ...ele,
-                label: action.str,
-              };
-            } else {
-              return ele;
-            }
-          }),
-        };
-      }
-      case "update_radio": {
-        const formfields = state.formFields.filter(
-          (field) => field.id === action.formID
-        )[0] as radioField;
-
-        return {
-          ...state,
-          formFields: state.formFields.map((ele) => {
-            if (
-              ele.id === action.formID &&
-              (ele.kind === "dropdown" ||
-                ele.kind === "radio" ||
-                ele.kind === "multiple")
-            ) {
-              return {
-                ...ele,
-                options: formfields.options.map((ele) => {
-                  if (ele.id === action.id) {
-                    return {
-                      ...ele,
-                      option: action.stOption,
-                    };
-                  } else {
-                    return ele;
-                  }
-                }),
-              };
-            } else {
-              return ele;
-            }
-          }),
-        };
-      }
-
-      default: {
-        return {
-          ...state,
-          formFields,
-        };
-      }
-    }
-  };
 
   const saveForm = () => {
     const prev_data = getLocalForms();
     state.id = Number(new Date());
     saveLocalForms([...prev_data, state]);
     navigate("/");
-  };
-
-  const getNewField = (formKind: formKinds, formLabel: string) => {
-    const textType: formKinds = "text";
-    const multiSelect: formKinds = "multiple";
-    switch (formKind) {
-      case "dropdown":
-        return {
-          kind: formKind,
-          id: Number(new Date()),
-          label: formLabel,
-          options: [],
-          value: "",
-        };
-
-      case "radio":
-        return {
-          kind: formKind,
-          id: Number(new Date()),
-          label: formLabel,
-          options: [],
-          value: "",
-        };
-
-      case "multi-select": {
-        return {
-          kind: multiSelect,
-          id: Number(new Date()),
-          label: formLabel,
-          options: [],
-          value: "",
-        };
-      }
-
-      case "text": {
-        return {
-          kind: formKind,
-          id: Number(new Date()),
-          label: formLabel,
-          type: formKind,
-          value: "",
-        };
-      }
-
-      case "textarea": {
-        return {
-          kind: formKind,
-          id: Number(new Date()),
-          label: formLabel,
-          value: "",
-        };
-      }
-
-      default: {
-        console.log("default");
-        return {
-          kind: textType,
-          id: Number(new Date()),
-          label: formLabel,
-          type: formKind,
-          value: "",
-        };
-      }
-    }
   };
 
   //Use effects
@@ -369,7 +99,7 @@ export default function Form(props: { formId: number }) {
   //State variables
 
   //Contains the current form data
-  const [state, dispatch] = useReducer(reducer, null, () => formData());
+  const [state, dispatch] = useReducer(reducerForm, null, () => formData());
 
   //Data corresponding to the fields
   const [newField, setNewField] = useState("");
